* Fixes for Microsoft Visual Studio 2015

+++ unizone/src/windows/wlaunchthread_win.cpp
@@ -19,7 +19,7 @@ WLaunchThread::GotoURL(const QString & url, int showcmd)
 	WString wurl(url);
 	PRINT("GotoURL: url = %S\n", wurl.getBuffer());
 
-    if ((long)ShellExecuteW(NULL, NULL, wurl.getBuffer(), NULL, NULL, showcmd) > 32)
+    if ((uintptr_t)ShellExecuteW(NULL, NULL, wurl.getBuffer(), NULL, NULL, showcmd) > 32)
         return true;
 
     return false;
